backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==1.67.0
google-cloud-firestore==2.19.0
python-multipart==0.0.12
websockets==13.1
httpx==0.28.1
python-dotenv==1.0.1
